RE-187 Update SHA's to include var override fixes
Currently the group_var based overrides do not work
due to role variables having a higher precedence.

The OSA roles have been fixed to ensure that group_var
override work. This updates all the SHA's to include
those fixes.

The Octavia SHA is updated to ensure that it is
compatible with the OSA SHA changes.