RE-187 Move AIO bootstrap group_var adjustments
If the group_vars for artifact implementation are not
moved, the bootstrap process puts the rpco repo down
when executing the pip_install role.

This patch ensures that the settings are removed prior
to the bootstrap to eliminate this happening.

(cherry picked from commit 6dd6827b376de9c6d817fba86aff24f1f2021771)