RE-155 Move rpco apt source file config copy var
With the variable set in the lxc_hosts group_var
it is still set when there are no apt artifacts
and the task fails due to the undefined
rpco_mirror_apt_filename variable.

This ensures that all the apt artifact config is
set in one file so that it is easy to remove
when necessary.