RE-204 Filter LXC/xenial packages in UCA/trusty
The LXC packages which are versioned 16.04 (Xenial)
in the UCA mitaka/trusty-updates repo are conflicting
with the backports packages.

As we're expecting packages for Trusty, not Xenial to
be used and hve always used the backports packages, we
will filter them from the repo when doing the repo
mirror into aptly. This prevents the conflict.