LA-444 Ensure online migrations run correctly
The online migrations are disabled when they should
have run once per software version. When the task that
sets data_migration is skipped, it is misunderstood as
a success and the data migration facts are changed.

This is contrary to the intent of the task set, so the
role updates included in this patch include fixes to
ensure that the task which disables the need for
migrations is only activated when it should be.