RO-2435 Remove max_overflow overrides
The max_overflow override sets the maximum number of simultaneous
connected database threads. SQLAlchemy recommends setting this to 10
and oslo.db accepts this default. OSA uses the default for most
services but it bumps the max to 20 for very busy services, such as
nova and neutron.

This patch removes the override since it is no longer necessary to
override the max_overflow value. Bugs existed in Icehouse/Kilo that
caused excessive database threads when lots of instances were being
built, but those bugs have been fixed for years.