Remove untested role from rpco
The openstack-ops role is untested and under active development. Until
this role can be stabalized, is no longer being force pushed against
and has some testing it should not be part of the rpco product
offering. If this role is being used by folks in the support
organization it can cloned and installed as needed outside the confinds
of the RPCO prodcut offering.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>