Remove unneeded overrides (#2516)
* RO-2429 Remove neutron_api_workers override

This patch removes the neutron_api_workers override that set the
number of API workers to 64. OSA already has code that calculates
an optimal number of neutron api workers by taking the number of
CPU cores and dividing by 2. If that value is incorrect, we should
change the logic in OSA.

* RO-2430 Remove rpc_thread_pool_size override

Building lots of VMs (> 64) at a time in Icehouse/Kilo lead to RPC
thread exhaustion, which means VMs never came online. That was fixed
in a patch to nova during the Kilo release and it's recommended to
use the default of 64 RPC threads. The oslo.messaging project still
recommends this default in Pike.

This patch removes the RPC thread pool overrides, which reduces the
load on RabbitMQ and saves memory.

Upstream fix: https://review.openstack.org/#/c/132202/
Upstream bug: https://bugs.launchpad.net/neutron/+bug/1372049

* RO-2431 Remove rpc_response_timeout overrides

In the Havana/Grizzly releases, RPC thread response timeout overrides
were required because of an issue with iptables when lots of
simultaneous builds were occurring. This was fixed in the Havana
release and these overrides are no longer needed.

The OSA default of 60 seconds is strongly recommended by the
oslo.messaging project. This patch removes the override so that the
safe default can be used.

Upstream Bug: https://bugs.launchpad.net/nova/+bug/1199433
Upstream Fix: https://review.openstack.org/#/c/40688/

* RO-2432 Remove db_max_pool_size overrides

This default in OSA for db pool size overrides has been 120 since
Newton. We don't need to overrid this in rpc-openstack.

This patch removes those overrides.

* RO-2434 Remove pool_timeout overrides

RPC-O overrides the pool_timeout to 60 seconds, wheras oslo.db,
SQLAlchemy, and OSA all recommend 30 seconds as a maximum. There
were some bugs around building lots of instances and hitting db pool
timeouts back in the Icehouse/Kilo timeframes, but these bugs have
been fixes for quite some time.

This patch removes the override and uses the OSA default of 30 seconds.
If we're really expecting to wait up to 60 seconds for a successful
database connection, then we're making some serious MariaDB/OpenStack
issues in a deployment.