RLM-207 Correct influxdb/grafana repo overrides
The data structure used in rpc-maas is different to
all the OSA repositories, so the standard data in
'rpco_apt_repo' does not work.

In this patch we correct the overrides to ensure
that the right data structure is provided to make
use of the apt artifacts when executing the rpc-maas
playbooks.

(cherry picked from commit 88ffdeacfe2c54878f8f8481fd48a1d97acd9031)