RO-2479 Correct git repo preparation for OSA
The current preparation process does not work when
executing it on an existing rpc-openstack checkout.
It tries to use a submodule folder moved to a
different location, then updating it as a normal
git repo. This process fails.

If the user then tries to switch back to a different
branch which does have the submodule, it no longer
works properly either, leaving both repositories
broken.

This patch simplifies the process and prevents it
making any changes or doing a git fetch if the SHA
for openstack-ansible is already correct.