RO-2540 Add OSA git checkout to py_pkgs lookup
In 4d12456 the use of a submodule was removed and
replaced with a symlink in the same location. The
py_pkgs lookup for the git artifact preparation
no longer works properly in this configuration as
the checkout is not in the expected place and/or
the lookup appears not to traverse the symlink.

This patch ensures that the /opt/openstack-ansible
folder is added to the list given to the lookup.
The path set is forced by the ansible bootstrap
script, regardless of where the rpc-openstack
checkout is.