RE-724 Implement pre_merge_test hook
This patch implements the RPC Release Engineering pre_merge_hook
as documented in [1] making use of the existing job definition
in the rpc-gating repo.

For the most part it intentionally does not change how the job
is executed - it is meant to be a simple transfer of all logic
from rpc-gating into the rpc-openstack repository so that the
RPC-O team can own the entire test execution in a branched
repository.

[1] https://rpc-openstack.atlassian.net/wiki/spaces/RE/pages/19005457/RE+for+Projects