RE-360 Include Reno notes in release notes
This changes updates the release notes that will be generated by the
rpc-gating generate_release_notes hook to include those managed by Reno.

The result of this change is that the release notes now returned will be
any Reno notes followed by those produced by rpc-differ. If there are no
Reno notes for the release that will be explicitly stated in the release
notes.

The code was restructured so that the code for generating the different
types of notes could be kept separate. A wrapper script is used to
create a single set of notes from the separate files.