RE-694 Replace artifacts if it is safe to do so
Currently the periodic job which builds artifacts is not
set to replace the artifacts every time it runs. Instead
of setting that on, therefore always replacing them, we
use a function to verify that the rpc_release version
set in the mainline branch differs from the value set in
the release candidate branch.

This protection is implemented to prevent artifacts from
being replaced by mistake once a new release candidate
branch has been cut, but the mainline branch has not had
the rpc_release value changed to the next version yet.

Using this method still allows the periodic job to be
executed manually with REPLACE_ARTIFACTS="YES", forcing
them to be replaced.

We only implement this check when PUSH_TO_MIRROR is already
set to "YES" when the script executes. This is only the
case in the periodic jobs which are the ones designed to
be run in sequence to to upload their changes.

(cherry picked from commit eb4284551ed5041c2f05c01767fb696e044eef31)