RE-1109 Implement branch overrides in PR tests
The artifact build PR tests which test the container
builds use the role developer_mod ability to build
the venv without using python artifacts. This is
necessary for a PR test when the PR includes a change
to rpc_release - reason being that for a new rpc_release
no job has been able to run to build the artifacts.

As newton reaches EOL, developer_mode no longer works
as it tries to make use of the stable/newton branch
which no longer exists.

In this patch we implement the overrides necessary to
make the PR tests use the newton-eol branch.