RE-1178 Remove initial rpc-differ reno dockerfile (#2664)
The initial rpc-differ which is run in the dockerfile
is not really necessary as the later execution will
pull down everything it needs.

Due to changes in the master repo, the execution of
rpc-differ is failing for the newton branch. This is
caused by the hard coded use of the master branch for
in the dockerfile.

As it's not needed, and we'd prefer not to have the
branches interfere with each other, we remove this
line.

A clone of the rpc-openstack repo is added to ensure
that osa-differ properly updates it. This is a work
around until the issue noted is resolved.

(cherry picked from commit 32d1a368a0479b155f9382af0c3d2bea1ed44b1d)