RO-3117 Split rsync_opts into list
The Ansible documentation specifies that the rsync_opts
should be an array, not a string. While this has worked
for us so far with Ansible 2.1.x it could break at any
time and definitely causes failure with Ansible 2.3.x
so we change this up to be implemented the correct way.