RO-3117 Revert no-owner/no-group for rsync
When setting owner/group to 'no' with the synchronise
module it passes the '--no-owner --no-group' parameters
to rsync which causes the '--chown' option to be ignored
and the ownership of the files to be set to the user
executing the rsync command (ie root). This results in
aptly not being able to operate on the database as the
files are owned by root instead of the aptly user.

When leaving the owner/group settings out of the module
settings no extra options are passed to rsync and '--chown'
works as expected.

This has been extensively validated to be a working
implementation for Ansible 2.1 and Ansible 2.2 in [1].

[1] https://gist.github.com/odyssey4me/5817160155cc9ff7bd46aaef87fc2c07