RO-3149 Update rpc_release file location
The dep_update script checks the value of the rpc_release
in the rc branch and the current branch, then does things
from there. Currently the location of the file is wrong,
so the script is failing.

Also, the release.yaml group_var set now includes two lines
containing rpc_release, so we ensure that the extraction
awk picks the right one.