[RO-3150] Fix partial py artifacts installation (#2697)
This change moves the pip install role to the configure-python-sources
task file. This is done so we're only ever executing it to setup our
hosts when our sources are present. This change also uses two facts when
running with artifacts. First the artifacts must be online and availanle.
Second the artifact source must be "enabled". This gives deployers the
ability to enable or disable artifacts as needed.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>