RE-1199 Cater for alpha/beta release versions
The dependency update script also handles the automated update of
rpc_release when it is appropriate to do so. Currently it only
does minor version increments. There is no way to cater to alpha
or beta releases.

Also, the dependency update script is currently not working correctly.
With the implementation of c4132cd the dependency update script was
adjusted to update playbooks/vars/rpc-release.yml but it did not remove
the adjustments to etc/openstack_deploy/group_vars/all/release.yml. The
resulting PR's from the changes are therefore malformed.

This patch implements changes to the script in order to cater for
alpha/beta releases according to the rpc_release naming scheme in
RE-1199. It will auto-increment both pre-release and minor versions
depending on the current value of rpc_release. This gives us more
flexibility to cater to releasing internally before doing customer
facing releases.

In order to begin consuming this immediately, the master branch release
version is set to r16.0.0-alpha.1 in this patch.