RE-1214 Use an appropriately sized data disk
Currently we choose the largest available secondary disk as the
data disk for an AIO build, however this may be too small to be
useful. This patch checks disqualifies disks that are smaller
than the minimum required size, allowing a fall back to using
the root disk if a suitable one is not found.

This ensures that disks which may be used for swap or other
purposes by the cloud provider are not considered a data disk.