RO-3239 Ensure that RPC_RELEASE is correct (#2726)
* RO-3239 Move get-rpc_release to scripts

In order to make get-rpc_release more useful generally,
rather than just gating, the script is moved into the
scripts directory.

* RO-3239 Ensure that RPC_RELEASE is correct

With the implementation of c4132cd the use of the
release.yml file from group_vars is no longer an
appropriate way to derive the RPC_RELEASE version.

This patch ensures that the RPC_RELEASE version is
correctly derived using a small python script in
order to read the release mapping.

* RO-3239 Use argparse for get-rpc_release.py

Given that the get-rpc_release script is now possibly
going to be used by humans, using argparse makes sense
to make it friendlier to use. This patch uses argparse
and adds some exception handling.

* RO-3239 Derive RPC_RELEASE correctly for imaging

The imaging process currently uses its own method to
derive the value for rpc_release. This method no longer
works since c4132cd merged.

Instead of uding its own method, given that it sources
the common functions, it already has an environment
variable available with the right version, so we use
that instead.