RO-3263 Do not configure apt sources again
When executing the configure-apt-sources playbook
it configures the apt sources on all hosts. There
is a repeat of this which uses the pip_install role
to reconfigure the apt sources again in the
configure-python-sources playbook. This repeat
setup is unnecessary.

Also, the use of the pip_install role to configure
the apt sources at this stage does not work, because
the group_vars put in place to set the right variables
are not loaded into memory as they've only just been
placed there during the same playbook execution run.
As such we end up with the UCA repo being placed by
pip_install instead of the RPC-O apt sources.

Given the two above reasons, this patch removes the
unnecessary role execution.