RLM-316 Update gating artifact option setting
In https://github.com/rcbops/rpc-gating/pull/586 the names
for the jobs were normalised to ensure more consistency.

This updates the script to ensure that the new values are
properly interpreted.

The package upgrade consistently uses 'upgrade' instead
of 'dist-upgrade' as using the latter requires a reboot
to make all core OS packages upgraded active. A reboot
cannot be facilitated through this script and upgrading
the extra packages is unnecessary given the frequent image
updates anyway.