RLM-917 Add ability to gate without artifacts
In [1] the ability to do periodic tests without artifacts
was implemented. This updates the gating script and
corresponding deploy script variables so that this test
does what it should be doing.

The environment variable for the apt artifact mode is
renamed to be closer to the Ansible variable, but the
RPC_ prefix is used to prevent any sort of namespace
clash.

[1] https://github.com/rcbops/rpc-gating/pull/586