RE-1267 Ensure host has python/python-yaml
In order to get the value for rpc_release, we need to use
a python script, so we need to make sure that python and
the python yaml library are installed. If not, the value
for rpc_release cannot be read and the check for the value
will result in a null return, causing all artifacts to be
disabled.