RO-3654 Update OSA SHA for apt-cacher-ng fix
With the repo cache enabled, while using container artifacts,
the apt-cacher-ng log folder is not setup after the initial
bind mount. As a result apt-cacher-ng fails to start. This
issue is fixed in https://review.openstack.org/532581.

As there have been no python package SHA bumps in newton, this
patch brings the OSA SHA forward to include the upstream patch.