RO-3316 Enable container image usage
With the containers builds fixed, they are now producing single
stream xz archives which are usable by machinectl when importing.

Without using the container images we would have to implement [1]
to downgrade any packages installed upstream in lxc-ci which are
not available in our apt repository, so it is better to re-enable
the use of container artifacts where this is done in CI rather
than in deployments.

[1] https://github.com/rcbops/rpc-artifacts/blob/b00e8059d6c969b176e5a5942a4034d57978433e/user_rcbops_artifacts_building.yml#L57-L96