RLM-3856 Set artifact variables in script
Artifact variables were getting set in the ENV file
but not inside the script so that when the deploy-infra1.sh
file was generated, it was actually getting the artifact
defaults instead of the overrides.  This ensures those
variables are set so that the generated deploy-infra1.sh
has the proper values.

(cherry picked from commit 643c42a0ad164dd6d5e652ba4971c9feb4fa8ffc)