RE-1304 Add workaround for nova-novncproxy fail
For all newton releases up to r14.8.0 when executing the os-nova-install.yml
playbook the ``nova-novncproxy`` and ``nova-spicehtml5proxy`` services will
fail. The workaround to resolve this issue is to restart the services.

This patch adds a release note to communicate the known issue. The fix will
be brought into the r14.9.0 release as part of the upstream dependency
updates.