RO-3937 Add ironic/nova to EOL repo list
Both ironic and nova have deleted the stable/newton branch
and need to now make use of the EOL tags. When building
the containers without the python artifacts we use developer
mode for the role execution, so we need to set the correct
var to ensure the EOL tag is used.