RO-3857 Add support for MNAIO log collection
This change replaces the rsync commands used to gather logs and
configuration files on an AIO with a playbook that can be used on both
AIOs and MNAIOs. For an AIO the playbook targets localhost and for an
MNAIO the playbook uses the MNAIO inventory host file created by
openstack-ansible-ops.

This cherry pick modifies `collect_logs.yml` to ensure the
Kibana/Selenium tests results are saved. The filename of the tempest
test results is also changed.

(cherry picked from commit a35a446935d93d6dbffae3f5fa65344873f23bc7)