RO-3973 Override MariaDB repo configuration
MariaDB released new packages yesterday which breaks MariaDB
clustering for multi-node deployments. This breaks non-artifacted
builds which use MariaDB 10.1.

As a workaround until an upstream fix is implemented, we
override the defaults set in OSA to ensure that we use the
last known good version of MariaDB.

As these settings are unnecessary when using apt artifacts,
the overrides are removed if apt artifacts are enabled so
that only the apt artifacts repo is used.