RE-36 Update freeze and thaw process
This commit updates gating/pre_merge_test/post_deploy.sh to include
the full gating-specified image name and scenario in the name used to
create the instance snapshot. This is necessary to differentiate the
different artifact modes and whether the scenario is swift, ironic, etc.

Additionally, we update gating/thaw/thaw.yml to use lxc-autostart to
containers so that the start order defined in each container's config
is honoured.