RO-3991 Rename OSA_RELEASE to OSA_RELEASE_BRANCH
`RPC_RELEASE` is set to the current version of rpc-openstack, e.g.
r16.1.0. Without this change `OSA_RELEASE` gets set to the OpenStack
release series, e.g. pike. This mismatch provides for a confusing user
experience.

This change renames `OSA_RELEASE` to `OSA_RELEASE_BRANCH` which should
prevent its use from being assumed to be equivalent to `RPC_RELEASE`.
This also allows `OSA_RELEASE` to be re-introduced in a later commit
with an equivalent use to `RPC_RELEASE`.