RO-3991 Extend get-rpc_release component support
`get-rpc_release.py` is extended by this change with an additional
command-line option to support retrieving the version of any component
of an RPCO release.

For example:

`playbooks/vars/rpc-release.yml` currently specifies the following for
pike.

```
pike:
  maas_release: 1.6.0
  osa_release: 00d84d91c488e4e92b3cc552bec6ebb9dd20e38d
  rpc_release: r16.1.0
```

With this change it is now possible to retrieve, not just the value of
`rpc_release` but also `osa_release` and `maas_release`.

`./scripts/get-rpc_release.py -f playbooks/vars/rpc-release.yml -c osa`