RO-3991 Set OSA_BRANCH for openstack-ansible-ops
This change fixes the use of multi-node AIO testing by ensuring that
when openstack-ansible-ops is used to prepare a host for deploying
openstack-ansible, the correct OSA SHA is used.

`openstack-ansible-ops/multi-node-aio/build.sh` requires `OSA_BRANCH` to
be set to override the default OSA branch, master, when configuring an
MNAIO for deploying OSA.

`OSA_RELEASE` is defined in `functions.sh` with the correct OSA ref for
the current checkout to follow the existing pattern established by
`RPC_RELEASE`. It is the used to set `OSA_BRANCH` which is required by
openstack-ansible-ops.

(cherry picked from commit 0bf27bcd226a5d67f01f47a5deb350b974e5916c)