(ASC-144) Source MNAIO_SSH from an exported file
This commit sources the `MNAIO_SSH` variable in the `run` script
from a file created by the `run_deploy_mnaio.sh`. This mechanism is
used to make the variable available to the parent shell in order to
pass it to the `run_system_tests.sh` script as expected.

Because the variable is created by a sub-shell to the one used by
the `run` command, exporting it does not make it available to the
parent shell. To get around this, the variable is stored to an
external file and sourced after the `run_deploy_mnaio.sh` script has
returned control to the parent.