(ASC-144) Ensure RE_HOOK dirs exist for system test
This commit adds `mkdir` commands for the `RE_HOOK_RESULT_DIR` and
the `RE_HOOK_RESULT_DIR` to ensure that they exist prior to collecting
the assets to move to them. The commands are followed by a `|| true`
conditional in order to suppress an `exit 1` in the event that the
directories have been created prior to these commands being executed.

Prior to this commit, collecting the JUnit XML results for system
testing failed because the `RE_HOOK_RESULT_DIR` location was properly
set but did not exist at the specified location path.