Fix for get-rpc_release.py ignoring RPC_PRODUCT_RELEASE
The scripts/get-rpc_release.py used by the deployment to allow
for what looks like branchless deployments was ignoring the
RPC_PRODUCT_RELEASE.  It looks like argparse pulls that
environment variable to use for a release_series variable.
A default of 'queens' was being set as well.  The EnvDefault
class being used by that argument only uses the environment
variable if no default is set.  This change is to allow any
setting of RPC_PRODUCT_RELEASE to override or set the default.

Issue: RO-4110