RO-4147 Use forked repo_build role for PBR fix
A new version of PBR was released which causes some services, like
gnocchi, to break in new and exciting ways. This change installs PBR
within the repo_build role which will ensure all of the correct tools
are available to the repo build process and are installed using known
constraints.