RE-65 Generate ARA report and gather results
The gate-check-commit script in OSA, which is used for deployment
execution in RPC-O queens & pike, installs ARA [1] which is a tool
that gathers the ansible runtime execution results for each task
and is able to generate html/junit result files.

In this patch we use ARA to generate an html report and a junit
report, and we gather the database and all results with the rest
of the log data.

This has proved very useful for diagnostics in OSA gating, so we
expect that it will be very useful to RPC-O too.

[1] https://github.com/openstack/ara

(cherry picked from commit d6266f50b6b471aee53ca7d15e467653cfab268b)