RO-4188 Update newton-rc rpc_release to r14.13.0
The current rpc_release value is r14.12.0 which was
recently released. The newton-rc value is wrong due
to the dependency update PR's not merging into
the newton branch in time for the last release, so
now we need to do this manually.

The OSA SHA is also updated to include the complete
release note which describes the appropriate
workaround for the resolveconf issue.