[FLEEK-36] Add neutron env to sandbox agents
The neutron agents in OSA master/queens have been sand-boxed. This process
installs the agents on-metal (the host) instead of building a container
for them which improves stability and simplifies connectivity to the
rest of the cloud. Privilege escalation will be better understood and
the risk of container restarts, outages, shutdowns, or human error will be
removed given that the service will be running on the physical machines.

The ability to sandbox these services has long been available through
the use of environment overrides. This changes makes use of the override
capabilities.

The sandbox capabilities are built into OSA's use of systemd which
provides much of the constructs a container provides without the hacks
required to specifically make neutron run within a container. This
change matches what the community is already doing and will ensure our
pike customers can upgrade to queens in a seamless manor without any
need for additional maintenance's.

Issue: https://rpc-openstack.atlassian.net/projects/FLEEK/issues/FLEEK-36
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>