Change container backend to machinectl
The gate has long been running containerized workloads using machinectl.
This change pulls in that option as our default. The change will reduce
the required storage used to run containerized workloads and improve
general image mangement.

An optional playbook has been added to assist deployers with general
host setup tasks. This playbook is not a required step and has been
documented in the README as such.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>