RO-4206 Split MNAIO execution to 3-stage playbooks
The current MNAIO playbook execution differs from the AIO
execution in that the MNAIO uses the 'setup-everything'
playbook vs the AIO using the 3-stage execution which is
the documented way of deploying.

Using the 3-stage execution results in the ARA reporting
showing the three distinct stages, and the resulting junit
file also showing them distinctly. Using this approach may
also reduce deploy host memory consumption during the
deployment which may contribute to improving the chances
of success.

As such, we change the deployment to use the 3-stage method
so that the execution style is more uniform and we get better
reporting.

(cherry picked from commit 3b41a42fbb907e9a9167c2c8883b50426297a4dd)