RO-4206 Use patched ansible fork for tests
In this patch we set the AIO/MNAIO builds to use a fork of Ansible
which exposes the apt errors so that we can diagnose the cause of
the apt fetch failures.

Backport note:
A different branch is used due to the differing Ansible versions
between Queens and Pike.

(cherry picked from commit aa31c3ecbfb87dd8024381e07303abe7d9376e8a)