RO-4222 force-overwrite when publishing snapshots
This comes up from time to time, and upon investigation each time
it turns out that the before/after md5sums are consistent. The
issue appears to be related to the use of a long-term host on
which all tests and publishing is executed.

Rather than re-architect the solution which is only used for newton
(and will retire with newton) we apply the force overwrite flag so
that this infrequent behaviour does not block PR's or releases.