RE-1501 Really, really generate the dstat report
The previous patch did not work, and was prematurely merged.
It did actually generate the report, but put it in the origin
place for the data after the log collection was done.

This patch changes the mechanism to use the RE_HOOK_ARTIFACT_DIR
this time for the source and resulting report so that it is
collected along with the rest of the data.

It also now will do a find for any number of dstat.csv files in
the RE_HOOK_ARTIFACT_DIR and do a report for all of them. This
means that if the MNAIO tooling makes use of dstat it should also
have a report generated and collected.

(cherry picked from commit 12901c75d5487007122d26f09d8606ebb3983eb9)