RO-4265 Update reno hook to support OSA fork
With newton upstream gone EOL, we need to use a fork
of the openstack/openstack-ansible repo. Until now
the rpc-differ/osa-differ repositories have not
supported doing this. This patch updates the versions
of the tooling used to support the option and changes
the hook to use the new CLI option.

(cherry picked from commit 9353eae4ea34ff1a1bd956213d26a6871bedc1fe)