RE-1466 Ensure the apt cache is updated
Before we can install the fontconfig package, we need to
update the apt cache so that apt is aware of where it
can be installed from.

This is required in order to make use of nodepool nodes
which do not come with an updated cache when handed over
to tests.