RE-1584 Change AIO hostname to 'aio1'
When using nodepool, the hostname of the node is very long, resulting
in the following inventory error:

Inventory script (.../dynamic_inventory.py) had an execution error:
The resulting combination of
[ "nodepool-rpco-14.2-xenial-base-pubcloud-dfw-0000019218"
+ "nova_scheduler_container" ] is longer than 52 characters. This
combination will result in a container name that is longer than the
maximum allowable hostname of 63 characters. Before this process can
continue please adjust the host entries in your "openstack_user_config.yml"
to use a short hostname. The recommended hostname length is < 20 characters
long.

This patch therefore does what's been done in later versions of OSA,
which is to set the hostname to 'aio1'.