ASC-655 Add sdqc action
This commit adds a conditional for the "sdqc" action. If this action is
used, then the `RE_JOB_BRANCH` is changed to `openstack-ops-only` in
order to trigger the checkout of that branch when the
`run_system_tests.sh` script is executed. The purpose of this is to run
the deployment and validation of
[openstack-ops](https://github.com/rsoprivatecloud/openstack-ops/) in
isolation from other system test suites. Currently, this is necessary
due to a conflict in the deployment ansible code.