Remove functions source in deploy.sh
Sourcing the functions script in `deploy.sh` may result in failure on a
deployment where the base OS has been installed using a minimal image.
The functions script will be sourced by the install script so any and
all dependencies will be taken care of in the order in which they're
used.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>