Prep MNAIO and make pre stage idempotent
When using nodepool, pip is not installed on the host.
This causes the 'pre' stage to fail to install virtualenv,
then (as usual) the pre stage repeats three times and fails
at a different point because the script is not idempotent.

We make the script more idempotent so that any failure
should be more consistent, and we ensure that we handle
the preparation of apt sources and the installation of
python/python-yaml/virtualenv.

Part of this includes ensuring that we use the
Rackspace apt mirrors. This is a necessary change
specifically for MNAIO tests where they are setup
to use archive.ubuntu.com which is slow and fails
often due to there being more hops to it.

JIRA: RE-1730