Check most recent tag if there is no RC branch
In order to prepare to remove the RC branch, some
of the checks which use it need to change to check
against the most recent tagged release instead.

The artifact building currently checks the RC branch
to validate whether it's safe to update artifacts or
not. The script is now changed to also check the most
recent tagged release if there is no RC branch.

The current dependency update script uses the RC branch
for checking whether to update the version-in-code, and
then to do the update. This script is now changed to
check the most recent tagged release is there is no RC
branch.

JIRA: RE-1739