Check most recent tag if there is no RC branch
The current dependency update script uses the RC branch
for checking whether to update the version-in-code, and
then to do the update. This script is now changed to
check the most recent tagged release if there is no RC
branch. We also remove the 'new method' vs 'old method'
for the RC branch version checks given that the transition
which required it has long passed, so that's just cruft.

JIRA: RE-1739