Reduce complexity of tox test execution
Currently the tox tests are executed via multiple layers
of scripts. This patch flattens it into the gating/check/run
script to make following the code simpler.

It also moves the pre-installation of distro and python
packages to the 'pre' stage to improve the chances of success,
given that the 'pre' script is run three times before failing.