Move MNAIO tempest run into run_deploy_mnaio.sh
Running this from the parent script requires exporting vars
to a file and importing them from the parent. This is not
necessary if we just run it from the same script/environment.

This also puts all the MNAIO bits in a single script, making
it slightly easier to follow what's going on.

(cherry picked from commit cc6a90516fe13979bf05e359130e6c2dd50a3287)